backend/requirements.txt
# Core Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# LlamaIndex core + graph
llama-index-core==0.12.3
llama-index-graph-stores-neo4j==0.3.2
llama-index-llms-anthropic==0.6.2
llama-index-embeddings-voyageai==0.2.1
llama-index-vector-stores-pinecone==0.4.2

# Pinecone
pinecone-client==5.0.1

# Anthropic
anthropic==0.42.0

# VoyageAI
voyageai==0.3.2

# Source ingestion
exa-py==1.5.0
firecrawl-py==1.5.0
unstructured[all-docs]==0.16.11
arxiv==2.1.3
youtube-transcript-api==0.6.3

# Graph analysis
networkx==3.4.2

# HTTP / async
httpx==0.28.1
aiofiles==24.1.0
sse-starlette==2.2.1

# Utilities
slugify==0.0.1
python-slugify==8.0.4
tenacity==9.0.0
structlog==24.4.0
